# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthStatus": {
        "properties": {
          "state": {
            "enum": [
              "guest",
              "active",
              "inactive"
            ],
            "title": "State",
            "type": "string"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "vip_active": {
            "default": false,
            "title": "Vip Active",
            "type": "boolean"
          },
          "vip_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vip Label"
          }
        },
        "required": [
          "state",
          "username",
          "vip_active",
          "vip_label"
        ],
        "title": "AuthStatus",
        "type": "object"
      },
      "AuthStatusRequest": {
        "properties": {
          "auth": {
            "discriminator": {
              "mapping": {
                "browser": "#/components/schemas/BrowserAuth-Input",
                "cookie_session": "#/components/schemas/CookieSessionAuth-Input",
                "guest": "#/components/schemas/GuestAuth-Input"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/GuestAuth-Input"
              },
              {
                "$ref": "#/components/schemas/BrowserAuth-Input"
              },
              {
                "$ref": "#/components/schemas/CookieSessionAuth-Input"
              }
            ],
            "title": "Auth"
          }
        },
        "title": "AuthStatusRequest",
        "type": "object"
      },
...
      "Body_create_cookie_session_api_auth_cookie_sessions_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_create_cookie_session_api_auth_cookie_sessions_post",
        "type": "object"
      },
      "BrowserAuth-Input": {
        "properties": {
          "browser": {
            "enum": [
              "chrome",
              "edge",
              "firefox"
            ],
            "title": "Browser",
            "type": "string"
          },
          "kind": {
            "const": "browser",
            "default": "browser",
            "title": "Kind",
            "type": "string"
          },
          "profile": {
            "anyOf": [
              {
                "maxLength": 260,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile"
          }
        },
        "required": [
          "browser"
        ],
        "title": "BrowserAuth",
        "type": "object"
      },
...
      "CookieSessionResult": {
        "properties": {
          "cookie_count": {
            "title": "Cookie Count",
            "type": "integer"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "cookie_count"
        ],
        "title": "CookieSessionResult",
        "type": "object"
      },
      "CreateJobRequest-Input": {
        "properties": {
          "audio_format": {
            "$ref": "#/components/schemas/AudioFormat",
            "default": "original"
          },
          "auth": {
            "discriminator": {
              "mapping": {
                "browser": "#/components/schemas/BrowserAuth-Input",
                "cookie_session": "#/components/schemas/CookieSessionAuth-Input",
                "guest": "#/components/schemas/GuestAuth-Input"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/GuestAuth-Input"
              },
              {
                "$ref": "#/components/schemas/BrowserAuth-Input"
              },
              {
                "$ref": "#/components/schemas/CookieSessionAuth-Input"
              }
            ],
            "title": "Auth"
          },
          "credential": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Credential",
            "type": "string"
          },
          "item_indices": {
            "items": {
              "type": "integer"
            },
            "maxItems": 100,
            "title": "Item Indices",
            "type": "array"
          },
          "item_urls": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Item Urls",
            "type": "array"
          },
          "media_kind": {
            "$ref": "#/components/schemas/MediaKind"
          },
          "output_dir": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Output Dir",
            "type": "string"
          },
          "page_indices": {
            "items": {
              "type": "integer"
            },
            "maxItems": 100,
            "title": "Page Indices",
            "type": "array"
          },
          "quality_height": {
            "anyOf": [
              {
                "maximum": 10000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Height"
          },
          "quality_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Id"
          },
          "video_mode": {
            "$ref": "#/components/schemas/VideoMode",
            "default": "compatible_mp4"
          }
        },
        "required": [
          "credential",
          "media_kind",
          "output_dir"
        ],
        "title": "CreateJobRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobView": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "item_results": {
            "items": {
              "$ref": "#/components/schemas/JobItemResult"
            },
            "title": "Item Results",
            "type": "array"
          },
          "progress": {
            "$ref": "#/components/schemas/JobProgress"
          },
          "request": {
            "$ref": "#/components/schemas/CreateJobRequest-Output"
          },
          "result_paths": {
            "items": {
              "type": "string"
            },
            "title": "Result Paths",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "request",
          "progress",
          "result_paths",
          "item_results",
          "error_code",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "JobView",
        "type": "object"
      },
...
      "OpenOutputRequest": {
        "properties": {
          "path": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "OpenOutputRequest",
        "type": "object"
      },
...
      "ResolveRequest": {
        "properties": {
          "auth": {
            "discriminator": {
              "mapping": {
                "browser": "#/components/schemas/BrowserAuth-Input",
                "cookie_session": "#/components/schemas/CookieSessionAuth-Input",
                "guest": "#/components/schemas/GuestAuth-Input"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/GuestAuth-Input"
              },
              {
                "$ref": "#/components/schemas/BrowserAuth-Input"
              },
              {
                "$ref": "#/components/schemas/CookieSessionAuth-Input"
              }
            ],
            "title": "Auth"
          },
          "credential": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Credential",
            "type": "string"
          }
        },
        "required": [
          "credential"
        ],
        "title": "ResolveRequest",
        "type": "object"
      },
...
      "ResolvedVideo": {
        "properties": {
          "aid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aid"
          },
          "bvid": {
            "title": "Bvid",
            "type": "string"
          },
          "canonical_url": {
            "title": "Canonical Url",
            "type": "string"
          },
          "duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/VideoPage"
            },
            "title": "Pages",
            "type": "array"
          },
          "selected_page": {
            "default": 1,
            "title": "Selected Page",
            "type": "integer"
          },
          "thumbnail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "uploader": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploader"
          }
        },
        "required": [
          "canonical_url",
          "bvid",
          "aid",
          "title",
          "uploader",
          "thumbnail",
          "duration",
          "selected_page",
          "pages"
        ],
        "title": "ResolvedVideo",
        "type": "object"
      },
...
      "ResourceResolveRequest": {
        "properties": {
          "auth": {
            "discriminator": {
              "mapping": {
                "browser": "#/components/schemas/BrowserAuth-Input",
                "cookie_session": "#/components/schemas/CookieSessionAuth-Input",
                "guest": "#/components/schemas/GuestAuth-Input"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/GuestAuth-Input"
              },
              {
                "$ref": "#/components/schemas/BrowserAuth-Input"
              },
              {
                "$ref": "#/components/schemas/CookieSessionAuth-Input"
              }
            ],
            "title": "Auth"
          },
          "credential": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Credential",
            "type": "string"
          }
        },
        "required": [
          "credential"
        ],
        "title": "ResourceResolveRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Bilidown",
    "version": "0.1.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/cookie-sessions": {
      "post": {
        "operationId": "create_cookie_session_api_auth_cookie_sessions_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_cookie_session_api_auth_cookie_sessions_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CookieSessionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Cookie Session"
      }
    },
    "/api/auth/cookie-sessions/{session_id}": {
      "delete": {
        "operationId": "delete_cookie_session_api_auth_cookie_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Cookie Session"
      }
    },
    "/api/auth/status": {
      "post": {
        "operationId": "auth_status_api_auth_status_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Auth Status"
      }
    },
    "/api/jobs": {
      "get": {
        "operationId": "list_jobs_api_jobs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobView"
                  },
                  "title": "Response List Jobs Api Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Jobs"
      },
      "post": {
        "operationId": "create_job_api_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job"
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job"
      }
    },
    "/api/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_job_api_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job"
      }
    },
    "/api/jobs/{job_id}/events": {
      "get": {
        "operationId": "job_events_api_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Events"
      }
    },
    "/api/jobs/{job_id}/retry": {
      "post": {
        "operationId": "retry_job_api_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Job"
      }
    },
    "/api/open-output": {
      "post": {
        "operationId": "open_output_api_open_output_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenOutputRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Open Output"
      }
    },
...
    "/api/resolve": {
      "post": {
        "operationId": "resolve_api_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvedVideo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve"
      }
    },
    "/api/resources/resolve": {
      "post": {
        "operationId": "resolve_resource_api_resources_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvedResource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Resource"
      }
    },
...
  }
}